linting and miscellaneous fixes
notably:
* adds `npm run lints` to auto-format then run two linters
* change the location of SMSInterface examples
