+++ src/app/sms-interfaces/sms-interfaces.spec.ts
@@ -6,15 +6,15 @@ describe('SMSDataItemInterface', () => {
   });
 
   it('should create a partly empty instance', () => {
-    let instance = new SMSDataItemInterface('timestamp');
+    const instance = new SMSDataItemInterface('timestamp');
     expect(instance).toBeTruthy();
     expect(instance.timestamp).toEqual('timestamp');
     expect(instance.name).toEqual(undefined);
   });
 
   it('should convert a partly empty instance to an array', () => {
-    let instance = new SMSDataItemInterface('timestamp');
-    let arr = instance.getFieldValues();
+    const instance = new SMSDataItemInterface('timestamp');
+    const arr = instance.getFieldValues();
     expect(arr.length).toEqual(7);
     expect(arr[0]).toEqual('timestamp');
     for (let i = 1; i < arr.length; i++) {
